user delete function added

+++ src/screens/UserListScreen.js
@@ -4,27 +4,25 @@ import { Table, Button } from 'react-bootstrap';
 import { FaTrash, FaEdit, FaCheck, FaTimes } from 'react-icons/fa';
 import { toast } from 'react-toastify';
 
-
 import Message from "../components/Message";
 import Loader from "../components/Loader";
-import {
-    useGetUsersQuery,
-} from '../slicers/usersApiSlice';
+import { useGetUsersQuery, useDeleteUserMutation
+    } from '../slicers/usersApiSlice';
 
 const UserListScreen = () => {
     const { data: users, refetch, isLoading, error } = useGetUsersQuery();
 
-    //const [deleteUser] = useDeleteUserMutation();
+    const [deleteUser] = useDeleteUserMutation();
 
     const deleteHandler = async (id) => {
-        /*if (window.confirm('Are you sure')) {
+        if (window.confirm('Are you sure')) {
             try {
                 await deleteUser(id);
                 refetch();
             } catch (err) {
                 toast.error(err?.data?.message || err.error);
             }
-        }*/
+        }
     };
 
     return(
